Update to 1.10.5 and ensure full compliance
As the patch submitter had been waiting for considerable time for merge,
I've merged their patch as is, and made necessary post adjustments for
merge into the repositories.

Notably the package is modified to use the correct paths, ypkg build steps,
and document that this is not entirely open source, as such for distribution
compliance we must ensure the binaries aren't modified (stripped) and ship
a copy of the unmodified license text with the package.

This fixes T452.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>